Refactor `SearchManagerPrinter` and `SearchManagerSummaryPrinter` to use `PrinterTxtTable` for table generation and improve code readability.

+++ engine-reports/src/main/java/net/chesstango/reports/engine/SearchManagerModel.java
@@ -27,7 +27,7 @@ public class SearchManagerModel implements Model<List<SearchResponse>> {
 
     String searchesName;
 
-    long searches;
+    int searches;
 
     int searchByOpenBookCounter;
     int searchByTreeCounter;
